chore: move selector package to inside of assertions module
